Filters follow session
Updated the getSessionUrl function in useVoting to include filter parameters in the URL
Modified the FilterPanel component to update the URL with filter changes
Updated the RestaurantFinder component to read filter parameters from the URL and apply them when loading a session
Now when you share a session link, it will include:
Price level filters
Distance filters
Rating filters
Cuisine type filters
When someone opens the shared link, they will see the same filters that were applied when the link was shared. This ensures that everyone in the session sees the same set of restaurants based on the shared filters.
The changes have been deployed successfully. You can test it by:
Setting some filters in your session
Copying the session link
Opening the link in a new browser window
Verifying that the filters are the same as in the original session